use docker buildx build to obtain image contents
Docker save copies all files from a container, which is not the same
as all files in an image. This is because the container image creates
things like /proc and /dev/console.

Use this truly atrocious hack of executing docker buildx as a subprocess
to get things going. Anything else seems like a massive complication since
OCI is such a gazonking mess to understand.