fix caching bug, again
The image cache was buggy since open file descriptor locks don't work
on directories. This was obscured by a buggy fileControl method, and so
the code behaved as if acquiring the lock always worked.

Fix this by switching back to flock. Deal with the non-atomiticy of
flock changes by revalidating that the contents directory doesn't
exist.