config: allow substituting tag via -vm.kernel
The configuration file specifies the default kernel image that is
used for a folder or repository. It's useful to use a different tag
to test a change on multiple versions. Right now this means repeating
the full URL.

Allow passing '-vm.kernel :tag' which substitutes 'tag' into the
default URL.