simplify invocation
Invoking vimto goes via the -exec flag:

    go test -exec vimto ...

This has the downside of running into the special rules that go test
has for arguments. For example, the following doesn't work:

    go test -exec vimto -vm.kernel /path -run Foo ./bar

go test stops parsing the command line args after -vm.kernel which leads
to very confusing behaviour. The fix is to always add arguments to the end
of the command line, but that is hard to remember and impossible to enforce
from our side. The user has to learn by error, which isn't great.

Additionally, there is no way to provide the user feedback as to what
vimto is doing, since go test buffers all output. For example, it might look
like the program is hanging while we are downloading a large OCI image.

Fix this by changing the way vimto is invoked. Instead of being called by go
test we call go test instead. The equivalent of the example above is now:

    vimto -kernel /path -- go test -run Foo ./bar

A nice side effect is that flags are less awkward to write out.