fix restoring tty settings on ctrl-c
qemu changes some settings when it detects that stdin is a TTY.
Specifically, it turns off echoing. On shutdown it restores the old
settings. However, this doesn't work when the process is interrupted
with ctrl-c because we react to os.Interrupt by killing the qemu
process. This means that there is no time for qemu to restore settings.
The result is a broken shell that you can type into, but which doesn't
echo anything.

Fix this by "forwarding" ctrl-c to qemu by sending it os.Interrupt and
giving it a grace time to shut down. This allows restoring the
terminal setting.