enable coredumps inside the vm
Enable collection of coredumps inside the VM. This can aid debugging by
using GOTRACEBACK:

    GOTRACEBACK=crash vimto -- go test

A panic in the testsuite (including a test timeout) then triggers a
coredump which can be inspected using delve.

The coredump is placed into the current working directory, alongside
with the test binary (if invoked via go test).