fix vimto exec for binaries in /tmp
vimto exec fails for binaries in /tmp, because I introduced a bug
into the shared directory code. This didn't trigger a failure in CI
because we use a TMPDIR which is a subdirectory of the current
working directory.

Add a script helper which allocates a new TMPDIR and fix the bug.