image: rework atomic cache commit
The image cache uses a "contents" directory to store the extracted
image data. This makes it so that a manually extracted image is not
a valid cache entry, and therefore the code has to deal with both
separately.

Rewrite the atomic cache commit so that it doesn't need the contents
directory anymore.