don't extract image for each script test
The script test runner sets a new TMPDIR for each script. This changes
the location of the cache. Hence every invocation of vimto does a new
fetch, which adds up to a bit of time.

Add an explicit test for OCI behaviour and reuse the cached kernel in
other tests.