make missing -kernel error message less confusing
Passing a non-existing local path to -kernel currently gives a confusing
error message. For example `-kernel ./foo/bar` causes:

     Error: exec: retrieve kernel from OCI image: [...]
         fetch image: get from remote: Get "https://./v2/":
         dial tcp: lookup . on 127.0.0.53:53: no such host

This is because we try to resolve the kernel as an OCI image if a local
file doesn't exist.

Make this a bit less confusing by returning a better error message in
the common case where someone uses a "rooted" path like
/foo, ./foo or ../foo.